Refactor traffic_control admin into a package
All traffic_control Django admin definitions were in one module. Split
the admin classes into a package that is split into modules the same
way as models are.

There's no actual changes other than moving the admin classes around.

Refs LIIK-112
