+++ traffic_control/admin/__init__.py
@@ -1,0 +1,34 @@
+# flake8: noqa
+from .audit_log import AuditLogHistoryAdmin
+from .barrier import (
+    BarrierPlanAdmin,
+    BarrierPlanFileInline,
+    BarrierRealAdmin,
+)
+from .mount import (
+    MountPlanAdmin,
+    MountPlanFileInline,
+    MountRealAdmin,
+    PortalTypeAdmin,
+)
+from .road_marking import (
+    RoadMarkingPlanAdmin,
+    RoadMarkingPlanFileInline,
+    RoadMarkingRealAdmin,
+)
+from .signpost import (
+    SignpostPlanAdmin,
+    SignpostPlanFileInline,
+    SignpostRealAdmin,
+)
+from .traffic_light import (
+    TrafficLightPlanAdmin,
+    TrafficLightPlanFileInline,
+    TrafficLightRealAdmin,
+)
+from .traffic_sign import (
+    OrderedTrafficSignRealInline,
+    TrafficSignPlanAdmin,
+    TrafficSignPlanFileInline,
+    TrafficSignRealAdmin,
+)
